% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_windows.R
\name{extract_windows}
\alias{extract_windows}
\title{Local window extraction to create training and testing data for spatialized classification.}
\usage{
extract_windows(x, y, window_dim = c(1, 1), poly_extract = "all",
  remove_outside_bbox_points = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{Raster*. A Raster* used to extract data from.}

\item{y}{Spatial*. A SpatialPolygonsDataFrame or SpatialPointsDataFrame object used to extract the data.}

\item{window_dim}{Numeric vector. Two-element vector representing the width and height in pixel units of the window.  Default is 1,1 (a single pixel).}

\item{poly_extract}{Logical. FALSE currently unsupported, leave as TRUE.}

\item{remove_outside_bbox_points}{Logical. Remove any points falling outside of the raster bbox.}

\item{verbose}{Logical. Verbose execution.  Default = FALSE.}
}
\description{
Local window extraction to create training and testing data for spatialized classification.
}
\details{
For each pixel falling within a polygon, or under a point, a window around that pixel as given by the window_dim
parameter is extracted.  The dataframe variable names are labelled VN, where N is the Nth element of the array surrounding
the central pixel, moving line by line, and then band by band.  The output object is a list containing three elements,
the "extracted_windows" for use with the classifier, the "variable_lookup" which links the variable names against
the local radius, azimuth and band, and "window_dim" which is the same as the input "window_dim". 

For large extractions, if a foreach engine is registered, the extraction will run in parallel, one polygon per worker.
}
\examples{
# Extract 5x5 windows from a set of polygons.
training_raster <- setMinMax(
brick(system.file("external/tahoe_highrez.tif", package="spatial.tools")))

training_polys <- readOGR(dsn=system.file("external"),
		layer="spatialized.classification.training_v01")

window_dim = c(5,5)
extracted_training_polys <- extract_windows(x=training_raster,y=training_polys,window_dim=window_dim)

import raster rgdal foreach
}
\author{
Jonathan A. Greenberg (\email{spatial.tools@estarcion.net})
}

