% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_spatialized_classification.R
\name{predict_spatialized_classification}
\alias{predict_spatialized_classification}
\title{import spatial.tools randomForestSRC}
\usage{
predict_spatialized_classification(x, sc, classonly = TRUE,
  parallelmode = "none", verbose = FALSE, debugmode = FALSE,
  cores = detectCores())
}
\description{
import spatial.tools randomForestSRC
}

