% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialized_classification_accuracy.R
\name{spatialized_classification_accuracy}
\alias{spatialized_classification_accuracy}
\title{import caret}
\usage{
spatialized_classification_accuracy(sc, testing_data)
}
\description{
import caret
}

